Make LayoutRenderer host agnostic and shared
